Just disconnect if hostname was invalid after the handshake
As it turns out I first have to go through the encryption process before
I'm allowed to send a disconnect message back to the client. As this is kind
of impossible as the actual minecraft server needs to do the encryption I'll
just break the connection. Because of this a lot of my other plans are not
really possible either, a few still are so these will be next to implement.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>